requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.11.1
boto3==1.23.4
botocore==1.26.4
certifi==2021.10.8
charset-normalizer==2.0.12
distlib==0.3.4
Django==3.2.7
django-rest-framework==0.1.0
django-storages==1.12.3
djangorestframework==3.13.1
filelock==3.4.0
idna==3.3
jmespath==1.0.0
MouseInfo==0.1.3
mysqlclient==2.1.0
oauthlib==3.2.0
Pillow==9.0.1
platformdirs==2.4.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2021.1
pywhatkit==5.3
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tweepy==4.10.0
urllib3==1.26.9
virtualenv==20.10.0
wikipedia==1.4.0
